Round a value to a defined number of significant digits printing out trailing zeros, if applicable

Round a value to a defined number of significant digits printing out trailing zeros, if applicable

# S3 method for data.frame
pretty_round(x, ...)

pretty_signif(x, ...)

# S3 method for data.frame
pretty_signif(x, ...)

# S3 method for default
pretty_signif(x, digits = 6, sci_range = 6, sci_sep = "e", ...)

Arguments

x

The number to round.

...

Arguments passed to methods.

digits

integer indicating the number of significant digits.

sci_range

integer (or Inf) indicating when to switch to scientific notation instead of floating point. Zero indicates always use scientific; Inf indicates to never use scientific notation; otherwise, scientific notation is used when abs(log10(x)) > sci_range.

sci_sep

The separator to use for scientific notation strings (typically this will be either "e" or "x10^" for computer- or human-readable output).

Value

A string with the value.

Details

Values that are not standard numbers like Inf, NA, and NaN are returned as "Inf", "NA", and NaN.

See also